controllers/requirements.txt
aniso8601==7.0.0
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
Pillow==6.1.0
protobuf==3.9.1
pytz==2019.2
pyzbar==0.1.8
six==1.12.0
Werkzeug==0.15.6
